requirements.txt
psycopg2==2.9.11
psycopg2-binary==2.9.10
python-dotenv==1.0.1